+++ tests/e2e/test_openapydantic.py
@@ -7,7 +7,11 @@ import openapydantic
 
 @pytest.mark.asyncio
 async def test_load_api_empty() -> None:
-    file_path = os.path.join(os.path.dirname(__file__), "fixture", "api-empty.yaml")
+    file_path = os.path.join(
+        os.path.dirname(__file__),
+        "fixture",
+        "api-empty.yaml",
+    )
     with pytest.raises(ValueError):
         await openapydantic.load_api(file_path=file_path)
 
@@ -15,7 +19,9 @@ async def test_load_api_empty() -> None:
 @pytest.mark.asyncio
 async def test_load_api_no_version() -> None:
     file_path = os.path.join(
-        os.path.dirname(__file__), "fixture", "api-no-version.yaml"
+        os.path.dirname(__file__),
+        "fixture",
+        "api-no-version.yaml",
     )
     with pytest.raises(ValueError):
         await openapydantic.load_api(file_path=file_path)
@@ -24,7 +30,9 @@ async def test_load_api_no_version() -> None:
 @pytest.mark.asyncio
 async def test_load_api_raise_unsupported_version() -> None:
     file_path = os.path.join(
-        os.path.dirname(__file__), "fixture", "api-unsupported-version.yaml"
+        os.path.dirname(__file__),
+        "fixture",
+        "api-unsupported-version.yaml",
     )
     with pytest.raises(NotImplementedError):
         await openapydantic.load_api(file_path=file_path)